requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
faiss-cpu
langgraph
requests
fastapi
uvicorn
upstash-redis
python-dotenv
SpeechRecognition
python-multipart
streamlit
pydub
google-genai
